+++ crates/forge/src/new.rs
@@ -7,6 +7,7 @@ use indoc::formatdoc;
 use scarb_api::version::scarb_version;
 use scarb_api::{ScarbCommand, ensure_scarb_available};
 use semver::Version;
+use shared::consts::FREE_RPC_PROVIDER_URL;
 use std::env;
 use std::fs::{self, OpenOptions};
 use std::io::Write;
@@ -153,7 +154,7 @@ fn create_snfoundry_manifest(path: &PathBuf) -> Result<()> {
         # and https://foundry-rs.github.io/starknet-foundry/projects/configuration.html for more information
 
         # [sncast.default]                                         # Define a profile name
-        # url = "<YOUR_RPC_PROVIDER>" # Url of the RPC provider
+        # url = "{default_rpc_url}" # Url of the RPC provider
         # accounts-file = "../account-file"                        # Path to the file with the account data
         # account = "mainuser"                                     # Account from `accounts_file` or default account file that will be used for the transactions
         # keystore = "~/keystore"                                  # Path to the keystore file
@@ -161,6 +162,7 @@ fn create_snfoundry_manifest(path: &PathBuf) -> Result<()> {
         # block-explorer = "StarkScan"                             # Block explorer service used to display links to transaction details
         # show-explorer-links = true                               # Print links pointing to pages with transaction details in the chosen block explorer
         "#,
+            default_rpc_url = FREE_RPC_PROVIDER_URL,
         },
     )?;
 
@@ -305,7 +307,7 @@ fn add_fork_config(document: &mut DocumentMut) -> Result<()> {
 
     let mut fork_table = Table::new();
     fork_table.insert("name", Item::Value(Value::from("SEPOLIA_LATEST")));
-    fork_table.insert("url", Item::Value(Value::from("<YOUR_RPC_PROVIDER>")));
+    fork_table.insert("url", Item::Value(Value::from(FREE_RPC_PROVIDER_URL)));
 
     let mut block_id_table = Table::new();
     block_id_table.insert("tag", Item::Value(Value::from("latest")));